+++ Mytube/src/main/java/com/flab/Mytube/dao/UserDao.java
@@ -1,4 +1,15 @@
 package com.flab.Mytube.dao;
 
-public interface UserDao {
+import com.flab.Mytube.vo.User;
+import org.apache.ibatis.session.SqlSession;
+
+public class UserDao {
+    private SqlSession sqlSession;
+    public void setSqlSession(SqlSession sqlSession){
+        this.sqlSession=sqlSession;
+    }
+
+    public User getUser(String userId){
+        return sqlSession.selectOne("org.mybatis.spring.sample.mapper.UserMapper.getUser", userId);
+    }
 }
